[PERFORMANCE] Pickle5 ConnectionWrapper for VectorEnv at least (2-5% speed up) (#582)
PEP 574 introduced a zero-copy pickling method that makes the pickling of Numpy arrays significantly faster. However, this feature is only supported on the latest Pickle protocol (5) which is not used by the default for multiprocessing (or even torch.multiprocessing, although it is often used in wrappers like Queues).
I ported and slightly modified the torch Connection wrapper code to habitat-lab and improved it by adding a pickle5 backport. The backport port brings this functionality back to Python<3.8 After using this PR, it should remove a copy from NumPy arrays.

Furthermore, this should not affect the pickling of any other parts of the code stack or any other multiprocessing code.

This speed up is likely to be a lot more on higher resolution outputs from habitat-sim as well as the sensor numbers increase.

This change should be fully backwards compatible and speed things up across the board without any code changes for the end user.