The Pick Task: a Base Case of Rearrangement (#660)
Agent (a Fetch robot) is spawned close to a receptacle (a table) that holds multiple objects (e.g. cracker box, bowl). The task for the robot is to pick up a target object with center-of-mass coordinates s_0 (provided in robot’s coordinate system) as efficiently as possible without excessive collisions. We aim for systematic generalization to new clutter layout on the receptacle, to new objects, and to new receptacles.